Editor: Move the panel visibility state from the edit-post to the editor package
