+++ docs/reference-guides/data/data-core-edit-post.md
@@ -144,6 +144,8 @@ Returns true if the template editing mode is enabled.
 
 ### isEditorPanelEnabled
 
+> **Deprecated**
+
 Returns true if the given panel is enabled, or false otherwise. Panels are enabled by default.
 
 _Parameters_
@@ -157,6 +159,8 @@ _Returns_
 
 ### isEditorPanelOpened
 
+> **Deprecated**
+
 Returns true if the given panel is open, or false otherwise. Panels are closed by default.
 
 _Parameters_
@@ -169,6 +173,8 @@ _Returns_
 -   `boolean`: Whether or not the panel is open.
 
 ### isEditorPanelRemoved
+
+> **Deprecated**
 
 Returns true if the given panel was programmatically removed, or false otherwise. All panels are not removed by default.
 
@@ -408,6 +414,8 @@ _Returns_
 
 ### removeEditorPanel
 
+> **Deprecated**
+
 Returns an action object used to remove a panel from the editor.
 
 _Parameters_
@@ -484,6 +492,8 @@ Action that toggles Distraction free mode. Distraction free mode expects there a
 
 ### toggleEditorPanelEnabled
 
+> **Deprecated**
+
 Returns an action object used to enable or disable a panel in the editor.
 
 _Parameters_
@@ -495,6 +505,8 @@ _Returns_
 -   `Object`: Action object.
 
 ### toggleEditorPanelOpened
+
+> **Deprecated**
 
 Opens a closed panel and closes an open panel.
