+++ packages/edit-post/src/components/sidebar/post-excerpt/index.js
@@ -6,13 +6,13 @@ import { PanelBody } from '@wordpress/components';
 import {
 	PostExcerpt as PostExcerptForm,
 	PostExcerptCheck,
+	store as editorStore,
 } from '@wordpress/editor';
 import { useDispatch, useSelect } from '@wordpress/data';
 
 /**
  * Internal dependencies
  */
-import { store as editPostStore } from '../../../store';
 import PluginPostExcerpt from '../plugin-post-excerpt';
 
 /**
@@ -23,7 +23,7 @@ const PANEL_NAME = 'post-excerpt';
 export default function PostExcerpt() {
 	const { isOpened, isEnabled } = useSelect( ( select ) => {
 		const { isEditorPanelOpened, isEditorPanelEnabled } =
-			select( editPostStore );
+			select( editorStore );
 
 		return {
 			isOpened: isEditorPanelOpened( PANEL_NAME ),
@@ -31,7 +31,7 @@ export default function PostExcerpt() {
 		};
 	}, [] );
 
-	const { toggleEditorPanelOpened } = useDispatch( editPostStore );
+	const { toggleEditorPanelOpened } = useDispatch( editorStore );
 	const toggleExcerptPanel = () => toggleEditorPanelOpened( PANEL_NAME );
 
 	if ( ! isEnabled ) {
