+++ packages/edit-post/src/components/sidebar/post-taxonomies/taxonomy-panel.js
@@ -3,11 +3,7 @@
  */
 import { PanelBody } from '@wordpress/components';
 import { useSelect, useDispatch } from '@wordpress/data';
-
-/**
- * Internal dependencies
- */
-import { store as editPostStore } from '../../../store';
+import { store as editorStore } from '@wordpress/editor';
 
 function TaxonomyPanel( { taxonomy, children } ) {
 	const slug = taxonomy?.slug;
@@ -15,7 +11,7 @@ function TaxonomyPanel( { taxonomy, children } ) {
 	const { isEnabled, isOpened } = useSelect(
 		( select ) => {
 			const { isEditorPanelEnabled, isEditorPanelOpened } =
-				select( editPostStore );
+				select( editorStore );
 			return {
 				isEnabled: slug ? isEditorPanelEnabled( panelName ) : false,
 				isOpened: slug ? isEditorPanelOpened( panelName ) : false,
@@ -23,7 +19,7 @@ function TaxonomyPanel( { taxonomy, children } ) {
 		},
 		[ panelName, slug ]
 	);
-	const { toggleEditorPanelOpened } = useDispatch( editPostStore );
+	const { toggleEditorPanelOpened } = useDispatch( editorStore );
 
 	if ( ! isEnabled ) {
 		return null;
