+++ packages/edit-post/src/store/test/actions.js
@@ -205,83 +205,6 @@ describe( 'actions', () => {
 		} );
 	} );
 
-	describe( 'toggleEditorPanelEnabled', () => {
-		it( 'toggles panels to be enabled and not enabled', () => {
-			// This will switch it off, since the default is on.
-			registry
-				.dispatch( editPostStore )
-				.toggleEditorPanelEnabled( 'control-panel' );
-
-			expect(
-				registry
-					.select( editPostStore )
-					.isEditorPanelEnabled( 'control-panel' )
-			).toBe( false );
-
-			// Also check that the `getPreference` selector includes panels.
-			expect(
-				registry.select( editPostStore ).getPreference( 'panels' )
-			).toEqual( {
-				'control-panel': {
-					enabled: false,
-				},
-			} );
-
-			// Switch it on again.
-			registry
-				.dispatch( editPostStore )
-				.toggleEditorPanelEnabled( 'control-panel' );
-
-			expect(
-				registry
-					.select( editPostStore )
-					.isEditorPanelEnabled( 'control-panel' )
-			).toBe( true );
-
-			expect(
-				registry.select( editPostStore ).getPreference( 'panels' )
-			).toEqual( {} );
-		} );
-	} );
-
-	describe( 'toggleEditorPanelOpened', () => {
-		it( 'toggles panels open and closed', () => {
-			// This will open it, since the default is closed.
-			registry
-				.dispatch( editPostStore )
-				.toggleEditorPanelOpened( 'control-panel' );
-
-			expect(
-				registry
-					.select( editPostStore )
-					.isEditorPanelOpened( 'control-panel' )
-			).toBe( true );
-
-			expect(
-				registry.select( editPostStore ).getPreference( 'panels' )
-			).toEqual( {
-				'control-panel': {
-					opened: true,
-				},
-			} );
-
-			// Close it.
-			registry
-				.dispatch( editPostStore )
-				.toggleEditorPanelOpened( 'control-panel' );
-
-			expect(
-				registry
-					.select( editPostStore )
-					.isEditorPanelOpened( 'control-panel' )
-			).toBe( false );
-
-			expect(
-				registry.select( editPostStore ).getPreference( 'panels' )
-			).toEqual( {} );
-		} );
-	} );
-
 	describe( 'updatePreferredStyleVariations', () => {
 		it( 'sets a preferred style variation for a block when a style name is passed', () => {
 			registry
