+++ packages/edit-post/src/store/test/selectors.js
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import deepFreeze from 'deep-freeze';
-
-/**
  * Internal dependencies
  */
 import {
@@ -11,32 +6,11 @@ import {
 	isSavingMetaBoxes,
 	getActiveMetaBoxLocations,
 	isMetaBoxLocationActive,
-	isEditorPanelRemoved,
 	isInserterOpened,
 	isListViewOpened,
 } from '../selectors';
 
 describe( 'selectors', () => {
-	describe( 'isEditorPanelRemoved', () => {
-		it( 'should return false by default', () => {
-			const state = deepFreeze( {
-				removedPanels: [],
-			} );
-
-			expect( isEditorPanelRemoved( state, 'post-status' ) ).toBe(
-				false
-			);
-		} );
-
-		it( 'should return true when panel was removed', () => {
-			const state = deepFreeze( {
-				removedPanels: [ 'post-status' ],
-			} );
-
-			expect( isEditorPanelRemoved( state, 'post-status' ) ).toBe( true );
-		} );
-	} );
-
 	describe( 'hasMetaBoxes', () => {
 		it( 'should return true if there are active meta boxes', () => {
 			const state = {
